requirements.txt
# 주식 추천 서비스 - Python 의존성
fastapi==0.115.0
uvicorn[standard]==0.32.0
jinja2==3.1.4
python-multipart==0.0.12

# 한국 주식 데이터
pykrx==1.2.55
pandas==2.2.2
numpy==2.1.3

# 뉴스/웹 스크래핑
feedparser==6.0.11
requests==2.32.3
beautifulsoup4==4.12.3
aiohttp==3.11.6

# 데이터 처리
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.6.1
